+++ source/extensions/filters/network/mysql_proxy/mysql_utils.cc
@@ -21,7 +21,7 @@ void BufferHelper::addUint32(Buffer::Instance& buffer, uint32_t val) {
 
 void BufferHelper::addString(Buffer::Instance& buffer, const std::string& str) { buffer.add(str); }
 
-std::string BufferHelper::encodeHdr(const std::string& cmd_str, int seq) {
+std::string BufferHelper::encodeHdr(const std::string& cmd_str, uint8_t seq) {
   MySQLCodec::MySQLHeader mysqlhdr;
   mysqlhdr.fields_.length_ = cmd_str.length();
   mysqlhdr.fields_.seq_ = seq;
@@ -147,7 +147,7 @@ int BufferHelper::peekUint32(Buffer::Instance& buffer, uint32_t& val) {
 
 void BufferHelper::consumeHdr(Buffer::Instance& buffer) { buffer.drain(sizeof(uint32_t)); }
 
-int BufferHelper::peekHdr(Buffer::Instance& buffer, int& len, int& seq) {
+int BufferHelper::peekHdr(Buffer::Instance& buffer, uint32_t& len, uint8_t& seq) {
   uint32_t val = 0;
   if (peekUint32(buffer, val) != MYSQL_SUCCESS) {
     return MYSQL_FAILURE;
